requirements.txt
# === Core Data Handling ===
pandas>=2.2.2
numpy>=2.0.2
pyarrow>=16.0.0
fastparquet>=2024.2.0
pathlib; python_version >= "3.8"

# === Binance API & Networking ===
python-binance>=1.0.19
requests>=2.32.3
websockets>=12.0

# === Machine Learning / Prophet ===
prophet>=1.1.7
cmdstanpy>=1.2.5
scikit-learn>=1.5.2
scipy>=1.13.1
statsmodels>=0.14.2

# === Visualization & Reporting ===
matplotlib>=3.10.0
seaborn>=0.13.2
plotly>=5.24.0

# === Web / API Server ===
Flask>=3.0.3
fastapi>=0.115.0
uvicorn>=0.29.0

# === Automation & Scheduling ===
schedule>=1.2.1

# === Utilities & System ===
tqdm>=4.67.1
importlib_resources>=6.5.2
logging; python_version >= "3.8"
typing_extensions>=4.12.2

# === Optional (for Cloud / Deployment) ===
gunicorn>=23.0.0
psutil>=6.1.0